Switch to dotenv for local development
I originally used commander for local development because I liked the UX of adding secrets as command line flags vs. having the secrets in my repo. Unfortunately, this made it cumbersome to pass the secrets around the application.

I've instead opted to use dotenv. Now, I add the secrets to a .env file locally, which is never committed to the repo. This ends up having a superior UX because secrets can now easily be accessed from any file in the app.