Remove gray from embed colors
I was using gray as a "default" color when no special color was needed, but it only looked good in dark mode. Thankfully, I realized I don't need to do this. If you don't override the embed color, Discord already defaults to gray - and it automatically works with both dark and light modes.